backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
pydantic-settings
python-dotenv
python-multipart
python-jose[cryptography]
passlib
